+++ lib/number/index.ts
@@ -1,0 +1,59 @@
+'use strict';
+
+import {isNumber}               from './is';
+import {isNumberAbove}          from './isAbove';
+import {isNumberAboveOrEqual}   from './isAboveOrEqual';
+import {isNumberBelow}          from './isBelow';
+import {isNumberBelowOrEqual}   from './isBelowOrEqual';
+import {isNumberBetween}        from './isBetween';
+import {isInteger}              from './isInteger';
+import {isIntegerAbove}         from './isIntegerAbove';
+import {isIntegerAboveOrEqual}  from './isIntegerAboveOrEqual';
+import {isIntegerBelow}         from './isIntegerBelow';
+import {isIntegerBelowOrEqual}  from './isIntegerBelowOrEqual';
+import {isIntegerBetween}       from './isIntegerBetween';
+import {isNumericalNaN}         from './isNumericalNaN';
+import {randomBetween}          from './randomBetween';
+import {randomIntBetween}       from './randomIntBetween';
+import {round}                  from './round';
+import {toPercentage}           from './toPercentage';
+
+export {
+    isNumber,
+    isNumber as is,
+    isNumber as isNum,
+    isNumberAbove,
+    isNumberAbove as isNumGt,
+    isNumberAbove as isGt,
+    isNumberAboveOrEqual,
+    isNumberAboveOrEqual as isNumGte,
+    isNumberAboveOrEqual as isGte,
+    isNumberBelow,
+    isNumberBelow as isNumLt,
+    isNumberBelow as isLt,
+    isNumberBelowOrEqual,
+    isNumberBelowOrEqual as isNumLte,
+    isNumberBelowOrEqual as isLte,
+    isNumberBetween,
+    isNumberBetween as isNumBetween,
+    isInteger,
+    isInteger as isInt,
+    isIntegerAbove,
+    isIntegerAbove as isIntGt,
+    isIntegerAboveOrEqual,
+    isIntegerAboveOrEqual as isIntGte,
+    isIntegerBelow,
+    isIntegerBelow as isIntLt,
+    isIntegerBelowOrEqual,
+    isIntegerBelowOrEqual as isIntLte,
+    isIntegerBetween,
+    isIntegerBetween as isIntBetween,
+    isNumericalNaN,
+    randomBetween,
+    randomBetween as randBetween,
+    randomIntBetween,
+    randomIntBetween as randIntBetween,
+    round,
+    toPercentage,
+    toPercentage as toPct,
+};